+++ src/modules/basic_support/unary_transforms_procedures.f90
@@ -69,6 +69,8 @@ module subroutine associateFunctionPointer(name,funcPointer)
         funcPointer => unarySuperbee
     case("minmodLimiter")
         funcPointer => unaryMinmod
+    case("absFloor")
+        funcPointer => absFloor
     case("none")
         funcPointer => null()
     case default 
@@ -384,6 +386,20 @@ pure module function unarySlopeRatio(input,realParams,intParams,logicalParams) r
         output = max(real(0,kind=rk),min(input,real(1,kind=rk)))
 
     end function unaryMinmod
+!-----------------------------------------------------------------------------------------------------------------------------------
+    pure module function absFloor(input,realParams,intParams,logicalParams) result(output)
+    !! Floor value filter unary wrapper, uses the absolute value of the first real param as the floor value.
+    !! The result it sign(input) * max(abs(input),abs(realParams(1)))
+
+        real(rk)               ,dimension(:) ,intent(in) :: input 
+        real(rk)     ,optional ,dimension(:) ,intent(in) :: realParams
+        integer(ik)  ,optional ,dimension(:) ,intent(in) :: intParams
+        logical      ,optional ,dimension(:) ,intent(in) :: logicalParams
+        real(rk) ,allocatable ,dimension(:)              :: output
+
+        output = sign(max(abs(input),abs(realParams(1))),input) 
+
+    end function absFloor
 !-----------------------------------------------------------------------------------------------------------------------------------
 end submodule unary_transforms_procedures
 !-----------------------------------------------------------------------------------------------------------------------------------