+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added option to explicitly change the maximum number of BDE integrator restarts (still hard-capped to 10)
 - Another timeloop mode has been added where the output points are set, and the code makes sure they coincide with integrator steps. If the output point is sufficiently far away, the standard timestep behaviour is recovered. 
 - It is now possible to set the initial output index. This is useful when restarting, allowing the user to avoid overwriting previous output files.
+- New unary transform for flooring variables
 
 ### Bug Fixes